Qt: move guest memory dump code to own class
