+++ rpcs3/rpcs3.vcxproj.filters
@@ -585,6 +585,12 @@
     <ClCompile Include="QTGeneratedFiles\Release\moc_anaglyph_settings_dialog.cpp">
       <Filter>Generated Files\Release</Filter>
     </ClCompile>
+    <ClCompile Include="QTGeneratedFiles\Debug\moc_guest_memory_dumper.cpp">
+      <Filter>Generated Files\Debug</Filter>
+    </ClCompile>
+    <ClCompile Include="QTGeneratedFiles\Release\moc_guest_memory_dumper.cpp">
+      <Filter>Generated Files\Release</Filter>
+    </ClCompile>
     <ClCompile Include="QTGeneratedFiles\Debug\moc_breakpoint_list.cpp">
       <Filter>Generated Files\Debug</Filter>
     </ClCompile>
@@ -1314,6 +1320,9 @@
     <ClCompile Include="rpcs3qt\anaglyph_settings_dialog.cpp">
       <Filter>Gui\settings</Filter>
     </ClCompile>
+    <ClCompile Include="rpcs3qt\guest_memory_dumper.cpp">
+      <Filter>Gui\utils</Filter>
+    </ClCompile>
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="Input\ds4_pad_handler.h">
@@ -1565,7 +1574,7 @@
     <ClInclude Include="rpcs3qt\steam_utils.h">
       <Filter>Gui\utils</Filter>
     </ClInclude>
-    <ClInclude Include="rpcs3qt\anaglyph_settings_dialog.h">
+    <ClInclude Include="rpcs3qt\render_creator.h">
       <Filter>Gui\settings</Filter>
     </ClInclude>
   </ItemGroup>
@@ -1783,9 +1792,6 @@
     <CustomBuild Include="rpcs3qt\microphone_creator.h">
       <Filter>Gui\settings</Filter>
     </CustomBuild>
-    <CustomBuild Include="rpcs3qt\render_creator.h">
-      <Filter>Gui\settings</Filter>
-    </CustomBuild>
     <CustomBuild Include="rpcs3qt\patch_manager_dialog.h">
       <Filter>Gui\patch manager</Filter>
     </CustomBuild>
@@ -1926,6 +1932,15 @@
     </CustomBuild>
     <CustomBuild Include="rpcs3qt\game_list_actions.h">
       <Filter>Gui\game list</Filter>
+    </CustomBuild>
+    <CustomBuild Include="rpcs3qt\content_integrity.h">
+      <Filter>Gui\game list</Filter>
+    </CustomBuild>
+    <CustomBuild Include="rpcs3qt\anaglyph_settings_dialog.h">
+      <Filter>Gui\settings</Filter>
+    </CustomBuild>
+    <CustomBuild Include="rpcs3qt\guest_memory_dumper.h">
+      <Filter>Gui\utils</Filter>
     </CustomBuild>
   </ItemGroup>
   <ItemGroup>
@@ -2081,10 +2096,5 @@
     <Text Include="CMakeLists.txt">
       <Filter>buildfiles\cmake</Filter>
     </Text>
-  </ItemGroup>
-  <ItemGroup>
-    <CustomBuild Include="rpcs3qt\content_integrity.h">
-      <Filter>Gui\game list</Filter>
-    </CustomBuild>
   </ItemGroup>
 </Project>
