+++ rpcs3/rpcs3qt/main_window.h
@@ -109,7 +109,6 @@ private Q_SLOTS:
 	void BootVSH();
 	void BootSavestate();
 	void BootRsxCapture(std::string path = "");
-	void DumpGuestMemory();
 	void DecryptSPRXLibraries();
 	void show_boot_error(game_boot_result status);
